basic

+++ src/main/java/com/example/redes_sociais/model_comentario/ComentRepository.java
@@ -3,5 +3,5 @@ package com.example.redes_sociais.model_comentario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ComentRepository extends JpaRepository<Comentario, Long> {
+public interface ComentRepository extends JpaRepository<Comentario, Integer> {
 }
